requirements.txt
# Python Requirements for Kafka Producer
# Install with: pip install -r requirements.txt

kafka-python==2.0.2
